requirements.txt
Django
gunicorn
psycopg[binary]
python-dotenv
django-extensions
django-template-partials
django-htmx
django-crispy-forms
crispy-bootstrap5